requirements.txt
django==2.2.2
djangorestframework==3.9.4
django-rest-swagger==2.2.0
psycopg2==2.8.2
pytest==4.6.3
factory-boy==2.12.0
pytest-django==3.5.0